Boiler Glow Removal
Boiler has gone through a lot of changes over the years as well as nerfs.

No longer a caste that would sit at the xenomorph backlines to bombard the marines, now it sits on the frontlines with an increased weakness to fire and a shorter view range, putting itself into the vision range of any other marines on the front as well as exposing itself to use its abilities now.

I believe the caste has earned the right to be free of the glow that it has been burdened with for years.